+++ core/src/main/java/org/jdbi/v3/core/array/SqlArrayTypes.java
@@ -23,14 +23,20 @@ import java.util.function.Function;
 import org.jdbi.v3.core.config.ConfigRegistry;
 import org.jdbi.v3.core.config.JdbiConfig;
 import org.jdbi.v3.core.enums.internal.EnumSqlArrayTypeFactory;
+import org.jdbi.v3.core.inference.JdbiInterceptorChain;
 import org.jdbi.v3.core.internal.JdbiOptionals;
 
 /**
  * Configuration class for SQL array binding and mapping.
  */
 public class SqlArrayTypes implements JdbiConfig<SqlArrayTypes> {
+
+    private final JdbiInterceptorChain<SqlArrayType<?>, SqlArrayTypeFactory> inferenceInterceptors = new JdbiInterceptorChain<>(
+        InferredSqlArrayTypeFactory::new);
+
     private final List<SqlArrayTypeFactory> factories = new CopyOnWriteArrayList<>();
     private SqlArrayArgumentStrategy argumentStrategy;
+
     private ConfigRegistry registry;
 
     public SqlArrayTypes() {
@@ -56,6 +62,7 @@ public class SqlArrayTypes implements JdbiConfig<SqlArrayTypes> {
     private SqlArrayTypes(SqlArrayTypes that) {
         factories.addAll(that.factories);
         argumentStrategy = that.argumentStrategy;
+        inferenceInterceptors.copy(that.inferenceInterceptors);
     }
 
     /**
@@ -100,7 +107,9 @@ public class SqlArrayTypes implements JdbiConfig<SqlArrayTypes> {
      * @throws UnsupportedOperationException if the argument is not a concretely parameterized type
      */
     public SqlArrayTypes register(SqlArrayType<?> arrayType) {
-        return register(new InferredSqlArrayTypeFactory(arrayType));
+        SqlArrayTypeFactory factory = inferenceInterceptors.process(arrayType);
+
+        return register(factory);
     }
 
     /**
@@ -130,6 +139,14 @@ public class SqlArrayTypes implements JdbiConfig<SqlArrayTypes> {
     @Override
     public void setRegistry(ConfigRegistry registry) {
         this.registry = registry;
+    }
+
+    /**
+     * Returns the {@link JdbiInterceptorChain} for the SqlArrayType inference. This chain allows registration of custom interceptors to change the standard
+     * type inference for the {@link SqlArrayTypes#register(SqlArrayType)} method.
+     */
+    public JdbiInterceptorChain<SqlArrayType<?>, SqlArrayTypeFactory> getInterceptorChain() {
+        return inferenceInterceptors;
     }
 
     @Override